Refactor use term views
