+++ ideax/ideax/views/__init__.py
@@ -5,8 +5,7 @@ import collections
 import mistune
 import csv
 
-from datetime import date
-from django.contrib.auth.decorators import login_required
+from django.contrib.auth.decorators import login_required, permission_required
 from django.shortcuts import render, get_object_or_404, redirect
 from django.utils import timezone
 from django.http import JsonResponse
@@ -16,7 +15,6 @@ from django.contrib import messages
 from django.utils.translation import ugettext_lazy as _
 from django.http import HttpResponse
 from django.conf import settings
-from django.contrib.auth.decorators import permission_required
 from django.core.files.storage import FileSystemStorage, default_storage
 from django.db import connection
 from django.http import StreamingHttpResponse
@@ -27,18 +25,19 @@ from martor.utils import LazyEncoder
 from ...users.models import UserProfile
 from ..models import (
     Idea, Criterion, Popular_Vote, Phase, Phase_History,
-    Category, Comment, Dimension, Evaluation, Category_Image, Use_Term, Challenge
+    Category, Comment, Dimension, Evaluation, Category_Image, Challenge
 )
-from ..forms import (
-    IdeaForm, CriterionForm,
-    EvaluationForm, ChallengeForm, UseTermForm, DimensionForm
-)
+from ..forms import IdeaForm, CriterionForm, EvaluationForm, ChallengeForm, DimensionForm
 from ...singleton import Profanity_Check
 from ...mail_util import MailUtil
 from ...util import get_ip, get_client_ip, audit
 
 from .category import category_edit, category_list, category_new, category_remove, get_category_list
 from .category_image import category_image_edit, category_image_list, category_image_new, category_image_remove
+from .use_term import (
+    get_term_of_user, get_use_term_list, accept_use_term, get_valid_use_term, save_use_term,
+    use_term_detail, use_term_edit, use_term_list, use_term_new, use_term_remove,
+)
 
 mail_util = MailUtil()
 
@@ -48,27 +47,6 @@ def index(request):
         audit(request.user.username, get_client_ip(request), 'LIST_IDEAS_PAGE', Idea.__name__, '')
         return idea_list(request)
     return render(request, 'ideax/index.html')
-
-
-@login_required
-def accept_use_term(request):
-    if not request.user.userprofile.use_term_accept:
-        user_profile = UserProfile.objects.get(user=request.user)
-        user_profile.use_term_accept = True
-        user_profile.acceptance_date = timezone.localtime(timezone.now())
-        user_profile.ip = get_ip(request)
-        user_profile.save()
-        messages.success(request, _('Term of use accepted!'))
-        # logger.info('%(username)s|%(ip_addr)s|%(message)s', {
-        # 'username': request.user.username, 'ip_addr': get_client_ip(request), 'message': 'Term of use accepted!'})
-    else:
-        messages.success(request, _('Term of use already accepted!'))
-        # logger.info('%(username)s|%(ip_addr)s|%(message)s', {
-        # 'username': request.user.username, 'ip_addr': get_client_ip(request),
-        # 'message': 'Term of use already accepted!'})
-
-    audit(request.user.username, get_client_ip(request), 'ACCEPT_TERMS_OF_USE_OPERATION', UserProfile.__name__, '')
-    return redirect('index')
 
 
 @login_required
@@ -500,7 +478,7 @@ def change_idea_phase(request, pk, new_phase):
                     [idea.author.user.email]
                 )
             )
-        except Exception as e:
+        except Exception:
             pass
 
     return redirect('index')
@@ -581,88 +559,6 @@ def idea_comments(request, pk):
                                          {"comments": Comment.objects.filter(idea__id=pk),
                                           "idea_id": pk})
     return JsonResponse(data)
-
-
-def get_term_of_user(request):
-    term = Use_Term.objects.filter(final_date__gte=timezone.now())
-    if term.exists():
-        return JsonResponse({"term": term[0].term})
-    else:
-        return JsonResponse({"term": _("No Term of Use found")})
-
-
-@login_required
-def use_term_list(request):
-    return render(request, 'ideax/use_term_list.html', get_use_term_list())
-
-
-def get_use_term_list():
-    return {'use_term_list': Use_Term.objects.all(), 'today': date.today()}
-
-
-@login_required
-@permission_required('ideax.add_use_term', raise_exception=True)
-def use_term_new(request):
-    if request.method == "POST":
-        form = UseTermForm(request.POST)
-    else:
-        form = UseTermForm()
-    return save_use_term(request, form, 'ideax/use_term_new.html', True)
-
-
-@login_required
-@permission_required('ideax.change_use_term', raise_exception=True)
-def use_term_edit(request, pk):
-    use_term = get_object_or_404(Use_Term, pk=pk)
-    if request.method == "POST":
-        use_term_form = UseTermForm(request.POST, instance=use_term)
-    else:
-        use_term_form = UseTermForm(instance=use_term)
-    return save_use_term(request, use_term_form, 'ideax/use_term_edit.html')
-
-
-def save_use_term(request, form, template_name, new=False):
-    if request.method == "POST":
-        if form.is_valid():
-            use_term = form.save(commit=False)
-            use_term.creator = UserProfile.objects.get(user=request.user)
-            if use_term.is_invalid_date():
-                messages.error(request, _('Invalid Final Date'))
-                return render(request, template_name, {'form': form})
-            if Use_Term.objects.getActive() and new:
-                messages.error(request, _('Already exists a active Term Of Use'))
-                return render(request, template_name, {'form': form})
-            use_term.save()
-            messages.success(request, _('Term of Use saved successfully!'))
-            return redirect('use_term_list')
-    else:
-        return render(request, template_name, {'form': form})
-
-
-@login_required
-@permission_required('ideax.delete_use_term', raise_exception=True)
-def use_term_remove(request, pk):
-    use_term = get_object_or_404(Use_Term, pk=pk)
-    if request.method == 'GET':
-        use_term.final_date = timezone.now()
-        use_term.save()
-        messages.success(request, _('Terms of Use removed successfully!'))
-        return render(request, 'ideax/use_term_list.html', get_use_term_list())
-
-
-@login_required
-def use_term_detail(request, pk):
-    use_term = get_object_or_404(Use_Term, pk=pk)
-    return render(request, 'ideax/use_term_detail.html', {'use_term': use_term})
-
-
-def get_valid_use_term(request):
-    use_terms = Use_Term.objects.all()
-    for term in use_terms:
-        if term.is_past_due:
-            valid_use_term = term
-            return render(request, 'ideax/use_term.html', {'use_term': valid_use_term})
-    return render(request, 'ideax/use_term.html', {'use_term': _("No Term of Use found")})
 
 
 def idea_detail_pdf(request, idea_id):
@@ -951,4 +847,7 @@ def dimension_remove(request, pk):
 __all__ = [
     'category_edit', 'category_list', 'category_new', 'category_remove', 'get_category_list',
     'category_image_edit', 'category_image_list', 'category_image_new', 'category_image_remove',
+    'get_term_of_user', 'get_use_term_list', 'accept_use_term',
+    'get_valid_use_term', 'save_use_term', 'use_term_detail', 'use_term_edit', 'use_term_list',
+    'use_term_new', 'use_term_remove',
 ]
